+++ src/eclipse/scaladoc/.classpath
@@ -3,8 +3,10 @@
 	<classpathentry kind="src" path="scaladoc"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/partest"/>
 	<classpathentry kind="var" path="SCALA_BASEDIR/lib/ant/ant.jar"/>
-	<classpathentry kind="con" path="org.scala-ide.sdt.launching.SCALA_COMPILER_CONTAINER"/>
-	<classpathentry kind="con" path="org.scala-ide.sdt.launching.SCALA_CONTAINER"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/scala-compiler"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/scala-library"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/scala-xml"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/scala-parser-combinators"/>
 	<classpathentry kind="output" path="build-quick-scaladoc"/>
 </classpath>
