+++ test/junit/scala/tools/nsc/symtab/SymbolTableForUnitTesting.scala
@@ -42,7 +42,7 @@ class SymbolTableForUnitTesting extends SymbolTable {
   object loaders extends symtab.SymbolLoaders {
     val symbolTable: SymbolTableForUnitTesting.this.type = SymbolTableForUnitTesting.this
     lazy val platform: symbolTable.platform.type = symbolTable.platform
-    protected def lookupMemberAtTyperPhaseIfPossible(sym: Symbol, name: Name): Symbol =
+    def lookupMemberAtTyperPhaseIfPossible(sym: Symbol, name: Name): Symbol =
       sym.info.member(name)
     protected override def compileLate(srcfile: AbstractFile): Unit =
       sys.error(s"We do not expect compileLate to be called in SymbolTableTest. The srcfile passed in is $srcfile")
